Python-Programs/Discord-Bot-Hikari-Lightbulb/requirements.txt
hikari
hikari-lightbulb
python_dotenv